+++ app/src/androidTest/java/woowacourse/movie/view/reservation/complete/ReservationCompleteActivityTest.kt
@@ -13,7 +13,7 @@ import org.junit.Test
 import woowacourse.movie.R
 import woowacourse.movie.model.reservation.ReservationInfo
 import woowacourse.movie.model.seat.Seats
-import woowacourse.movie.view.Extras
+import woowacourse.movie.view.util.Extras
 import java.time.LocalDate
 
 class ReservationCompleteActivityTest {
