--- app/src/main/java/woowacourse/movie/util/ExceptionMessages.kt
@@ -1,5 +1,0 @@
-package woowacourse.movie.util
-
-object ExceptionMessages {
-    const val FRAGMENT_BINDING_STATE_EXCEPTION = "바인딩이 초기화되지 않았습니다. onCreateView ~ onDestroyView 사이에만 사용해주세요."
-}
