+++ app/src/main/java/woowacourse/movie/view/main/reservationlist/ReservationListFragment.kt
@@ -7,8 +7,8 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import woowacourse.movie.databinding.FragmentReservationListBinding
-import woowacourse.movie.util.ExceptionMessages
 import woowacourse.movie.view.main.reservationlist.adapter.ReservationListAdapter
+import woowacourse.movie.view.util.ExceptionMessages
 
 class ReservationListFragment :
     Fragment(),
