+++ app/src/main/java/woowacourse/movie/view/util/Extras.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.util
 
 object Extras {
     object MovieData {
