+++ app/src/main/java/woowacourse/movie/view/util/ReservationUiFormatter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.util
 
 import android.icu.text.DecimalFormat
 import java.time.LocalDate
